refactor: move Result class to libs folder
