+++ app/api/templates.v2/database/MongoTemplatesDataSource.ts
@@ -12,7 +12,7 @@ import {
   DefaultTemplateNotFoundError,
   TemplateDoesNotExistError,
 } from 'api/core/domain/template/errors';
-import { Result, ResultType } from 'api/core/application/common/Result';
+import { Result, ResultType } from 'api/core/libs/Result';
 import { TemplatesDataSource } from '../contracts/TemplatesDataSource';
 import { Property } from '../model/Property';
 import { RelationshipProperty } from '../model/RelationshipProperty';
